+++ Lab02Tasks/Task06.c
@@ -1,0 +1,9 @@
+#include <stdio.h>
+
+int main() {
+
+    char ch[10] = "LMX";
+    printf("The reverse of %s is %c%c%c \n",ch, ch[2], ch[1], ch[0]);
+
+    return 0;
+}
