+++ Lab02Tasks/Task08.c
@@ -1,0 +1,10 @@
+#include <stdio.h>
+
+int main(){
+    float price = 5.99;
+    int qty = 4;
+    float total = price * qty;
+
+    printf("Total price: %.2f\n", total);
+    return 0;
+}
