--- Task01.c
@@ -1,10 +1,0 @@
-#include <stdio.h>
-
-int main(){
-    int id = 101;
-    char section = 'B';
-    float marks = 89.5;
-
-    printf("ID %d, Section %c Marks %.4f\n", id, section, marks);
-    return 0;
-}
