--- Task05.c
@@ -1,11 +1,0 @@
-#include <stdio.h>
-
-int main() {
-    float length = 25.12428;
-    float width = 12.59;
-
-    float area = length * width;
-
-    printf("Area = %.4f\n",area);
-
-}
